+++ src/web_client/views.rs
@@ -2,12 +2,13 @@ use std::path::Path;
 
 use actix_files::{Files, NamedFile};
 use actix_web::{
+    dev::{fn_service, ServiceRequest, ServiceResponse},
     guard,
+    http::header as http_header,
     web,
+    web::Data,
     HttpResponse,
     Resource,
-    dev::{fn_service, ServiceRequest, ServiceResponse},
-    web::Data,
 };
 use uuid::Uuid;
 
@@ -45,6 +46,23 @@ pub fn static_service(web_client_dir: &Path) -> Files {
         }))
 }
 
+fn activitypub_guard() -> impl guard::Guard {
+    guard::fn_guard(|ctx| {
+        is_activitypub_request(ctx.head().headers())
+    })
+}
+
+fn opengraph_guard() -> impl guard::Guard {
+    guard::fn_guard(|ctx| {
+        let headers = ctx.head().headers();
+        let maybe_user_agent = headers.get(http_header::USER_AGENT)
+            .and_then(|value| value.to_str().ok());
+        if let Some(user_agent) = maybe_user_agent {
+            user_agent == "Synapse (bot; +https://github.com/matrix-org/synapse)"
+        } else { false }
+    })
+}
+
 // DEPRECATED
 async fn profile_page_redirect_view(
     config: web::Data<Config>,
@@ -62,9 +80,7 @@ async fn profile_page_redirect_view(
 
 pub fn profile_page_redirect() -> Resource {
     web::resource("/profile/{profile_id}")
-        .guard(guard::fn_guard(|ctx| {
-            is_activitypub_request(ctx.head().headers())
-        }))
+        .guard(activitypub_guard())
         .route(web::get().to(profile_page_redirect_view))
 }
 
@@ -84,9 +100,7 @@ async fn profile_acct_page_redirect_view(
 
 pub fn profile_acct_page_redirect() -> Resource {
     web::resource("/@{acct}")
-        .guard(guard::fn_guard(|ctx| {
-            is_activitypub_request(ctx.head().headers())
-        }))
+        .guard(activitypub_guard())
         .route(web::get().to(profile_acct_page_redirect_view))
 }
 
@@ -104,10 +118,28 @@ async fn post_page_redirect_view(
     Ok(response)
 }
 
+async fn post_page_opengraph_view(
+    config: web::Data<Config>,
+    db_pool: web::Data<DbPool>,
+    post_id: web::Path<Uuid>,
+) -> Result<HttpResponse, HttpError> {
+    let db_client = &**get_database_client(&db_pool).await?;
+    let post = get_post_by_id(db_client, &post_id).await?;
+    let page = format!(
+        include_str!("opengraph.html"),
+        acct=post.author.acct,
+        instance_url=config.instance_url(),
+        image_path="/ogp-image.png",
+    );
+    let response = HttpResponse::Ok()
+        .content_type("text/html")
+        .body(page);
+    Ok(response)
+}
+
 pub fn post_page_redirect() -> Resource {
     web::resource("/post/{object_id}")
-        .guard(guard::fn_guard(|ctx| {
-            is_activitypub_request(ctx.head().headers())
-        }))
-        .route(web::get().to(post_page_redirect_view))
+        .guard(guard::Any(activitypub_guard()).or(opengraph_guard()))
+        .route(web::get().guard(activitypub_guard()).to(post_page_redirect_view))
+        .route(web::get().guard(opengraph_guard()).to(post_page_opengraph_view))
 }